schemas: widen LIST.maxEntries uint32 -> uint256 (ADR-0047)
uint32 caps a list at 4.29B members — an IPv4-class ceiling that real
planet-scale lists exceed (Asia ~4.7B, world ~8.1B population). Pre-data
devnet is the only free time to fix it.

Cost is zero: EAS attestation data is ABI-encoded into 32-byte words, so
uint32 and uint256 both occupy one slot — the LIST record stays 160 bytes.
Resolver addresses are nonce-CREATE (bytecode-independent), so only the LIST
schema UID changes; addresses and LIST_ENTRY_SCHEMA_UID are unchanged.

- Contracts: ListResolver / ListEntryResolver / ListReader / IListReader decode
  + cache + event + ListMode struct widened to uint256.
- Deploy: LIST_DEFINITION field string updated; freeze invariant check passes.
- Tests: schema string + AbiCoder types updated; 57 passing.
- Frontend: create forms parse maxEntries as BigInt and drop the uint32
  ceiling; reads render it as bigint (viem decodes uint256 as bigint).
- deployedContracts.ts regenerated against the pinned fork — ABI-only diff
  (uint32->uint256 x4), addresses byte-identical.
- ADR-0047 records the decision + the count-vs-discriminator sizing principle
  for the schema-freeze audit; ADR-0044 status + specs cross-referenced.

Permanence-tier: Etched (LIST schema field string -> new LIST_SCHEMA_UID)
Refs: ADR-0047, ADR-0044, ADR-0037

+++ packages/hardhat/contracts/ListEntryResolver.sol
@@ -82,7 +82,7 @@ contract ListEntryResolver is SchemaResolver {
         bool appendOnly;
         uint8 targetType;
         bytes32 targetSchema;
-        uint32 maxEntries;
+        uint256 maxEntries;
     }
 
     // LIST is non-revocable → cache is valid forever after first touch
@@ -144,7 +144,7 @@ contract ListEntryResolver is SchemaResolver {
             Attestation memory L = _eas.getAttestation(listUID);
             if (L.schema != LIST_SCHEMA_UID) revert NotAList();
             (d.allowsDuplicates, d.appendOnly, d.targetType, d.targetSchema, d.maxEntries) =
-                abi.decode(L.data, (bool, bool, uint8, bytes32, uint32));
+                abi.decode(L.data, (bool, bool, uint8, bytes32, uint256));
             d.exists = true;
             _decl[listUID] = d;
         }
